#25184 refactoring Config
