+++ dotCMS/src/main/java/com/dotmarketing/util/RuntimeConfigWriter.java
@@ -17,7 +17,7 @@ public class RuntimeConfigWriter {
     @Inject
     SystemTableConfigSource systemTableConfigSource;
 
-
+    // todo: here we have to handle the recursive system table config source call
     public void setOverride(final String key, final String value) {
 
         if (value==null) {
